requirements.txt
amqp==5.0.9
asgiref==3.5.0
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==37.0.2
decorator==5.1.1
Django==4.0.2
djangorestframework==3.13.1
dnspython==2.2.1
email-validate==1.1.2
email-validator==1.1.3
filelock==3.6.0
flake8==4.0.1
gunicorn==20.1.0
h11==0.13.0
idna==2.10
IMAPClient==2.2.0
iniconfig==1.1.1
kombu==5.2.3
mail-parser==3.15.0
mccabe==0.6.1
mypy==0.931
mypy-extensions==0.4.3
numpy==1.22.3
outcome==1.1.0
packaging==21.3
pandas==1.4.2
pluggy==1.0.0
prompt-toolkit==3.0.26
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.7
PySocks==1.7.1
pytest==7.0.0
pytest-django==4.5.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2021.3
PyYAML==6.0
requests==2.27.1
selenium==4.1.5
simplejson==3.17.6
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.35
sqlparse==0.4.2
tomli==2.0.0
trio==0.20.0
trio-websocket==0.9.2
typing_extensions==4.0.1
urllib3==1.26.8
validators==0.19.0
vine==5.0.0
wcwidth==0.2.5
wsproto==1.1.0
lxml
